pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diffusion_microscopy"
version = "0.1.0"
description = "A deep learning pipeline for generating microscopy images using diffusion models"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "pandas>=1.3.0",
    "scikit-image>=0.18.0",
    "tqdm>=4.62.0",
    "jupyter>=1.0.0",
    "pytest>=6.2.5",
    "black>=21.7b0",
    "flake8>=3.9.2",
    "pillow>=8.3.0",
    "albumentations>=1.0.3",
    "tensorboard>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.7b0",
    "flake8>=3.9.2",
    "mypy>=0.910",
    "isort>=5.9.3",
]

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 